backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart

openai
faiss-cpu

supabase
httpx

pydantic
pytest
pytest-cov

sqlalchemy
psycopg2-binary
email-validator

pypdf